+++ tests/integration/upsertGraph.js
@@ -163,10 +163,10 @@ module.exports = session => {
                     'select "Model1"."model1Id", "Model1"."id" from "Model1" where "Model1"."id" in (2) order by "Model1"."id" asc',
                     'select "Model1"."id" from "Model1" where "Model1"."id" in (3) order by "Model1"."id" asc',
                     'select "model2"."model1_id", "model2"."id_col" from "model2" where "model2"."model1_id" in (2) order by "model2"."id_col" asc',
-                    'select "Model1Model2"."model2Id" as "objectiontmpjoin0", "Model1"."id" from "Model1" inner join "Model1Model2" as "Model1Model2" on "Model1"."id" = "Model1Model2"."model1Id" where "Model1Model2"."model2Id" in (1, 2) order by "Model1"."id" asc',
+                    'select "Model1Model2"."model2Id" as "objectiontmpjoin0", "Model1"."id" from "Model1" inner join "Model1Model2" on "Model1"."id" = "Model1Model2"."model1Id" where "Model1Model2"."model2Id" in (1, 2) order by "Model1"."id" asc',
 
                     'delete from "model2" where "model2"."id_col" in (2) and "model2"."model1_id" in (2)',
-                    'delete from "Model1" where "Model1"."id" in (select "Model1"."id" from "Model1" inner join "Model1Model2" as "Model1Model2" on "Model1"."id" = "Model1Model2"."model1Id" where "Model1Model2"."model2Id" in (1) and "Model1"."id" in (5) order by "Model1"."id" asc)',
+                    'delete from "Model1" where "Model1"."id" in (select "Model1"."id" from "Model1" inner join "Model1Model2" on "Model1"."id" = "Model1Model2"."model1Id" where "Model1Model2"."model2Id" in (1) and "Model1"."id" in (5) order by "Model1"."id" asc)',
 
                     'insert into "Model1" ("model1Prop1") values (\'inserted manyToMany\') returning "id"',
                     'insert into "model2" ("model1_id", "model2_prop1") values (2, \'inserted hasMany\') returning "id_col"',
@@ -174,7 +174,7 @@ module.exports = session => {
 
                     'update "Model1" set "model1Prop1" = \'updated belongsToOne\' where "Model1"."id" = 3 and "Model1"."id" in (3)',
                     'update "model2" set "model2_prop1" = \'updated hasMany 1\', "model1_id" = 2 where "model2"."id_col" = 1 and "model2"."model1_id" in (2)',
-                    'update "Model1" set "model1Prop1" = \'updated manyToMany 1\' where "Model1"."id" in (select "Model1"."id" from "Model1" inner join "Model1Model2" as "Model1Model2" on "Model1"."id" = "Model1Model2"."model1Id" where "Model1Model2"."model2Id" in (1) and "Model1"."id" = \'4\' order by "Model1"."id" asc)'
+                    'update "Model1" set "model1Prop1" = \'updated manyToMany 1\' where "Model1"."id" in (select "Model1"."id" from "Model1" inner join "Model1Model2" on "Model1"."id" = "Model1Model2"."model1Id" where "Model1Model2"."model2Id" in (1) and "Model1"."id" = \'4\' order by "Model1"."id" asc)'
                   ]);
               }
 
@@ -603,18 +603,18 @@ module.exports = session => {
                     'select "Model1"."model1Id", "Model1"."id" from "Model1" where "Model1"."id" in (2) order by "Model1"."id" asc',
                     'select "Model1"."id" from "Model1" where "Model1"."id" in (3) order by "Model1"."id" asc',
                     'select "model2"."model1_id", "model2"."id_col" from "model2" where "model2"."model1_id" in (2) order by "model2"."id_col" asc',
-                    'select "Model1Model2"."model2Id" as "objectiontmpjoin0", "Model1"."id" from "Model1" inner join "Model1Model2" as "Model1Model2" on "Model1"."id" = "Model1Model2"."model1Id" where "Model1Model2"."model2Id" in (1, 2) order by "Model1"."id" asc',
+                    'select "Model1Model2"."model2Id" as "objectiontmpjoin0", "Model1"."id" from "Model1" inner join "Model1Model2" on "Model1"."id" = "Model1Model2"."model1Id" where "Model1Model2"."model2Id" in (1, 2) order by "Model1"."id" asc',
 
                     'update "Model1" set "model1Id" = NULL where "Model1"."id" = 2',
                     'update "model2" set "model1_id" = NULL where "model2"."id_col" in (2) and "model2"."model1_id" = 2',
-                    'delete from "Model1Model2" where "Model1Model2"."model1Id" in (select "Model1"."id" from "Model1" inner join "Model1Model2" as "Model1Model2" on "Model1"."id" = "Model1Model2"."model1Id" where "Model1Model2"."model2Id" in (1) and "Model1"."id" in (5) order by "Model1"."id" asc) and "Model1Model2"."model2Id" = 1',
+                    'delete from "Model1Model2" where "Model1Model2"."model1Id" in (select "Model1"."id" from "Model1" inner join "Model1Model2" on "Model1"."id" = "Model1Model2"."model1Id" where "Model1Model2"."model2Id" in (1) and "Model1"."id" in (5) order by "Model1"."id" asc) and "Model1Model2"."model2Id" = 1',
 
                     'insert into "Model1" ("model1Prop1") values (\'inserted manyToMany\') returning "id"',
                     'insert into "model2" ("model1_id", "model2_prop1") values (2, \'inserted hasMany\') returning "id_col"',
                     'insert into "Model1Model2" ("model1Id", "model2Id") values (8, 1), (6, 1) returning "model1Id"',
 
                     'update "Model1" set "model1Prop1" = \'updated root 2\', "model1Id" = NULL where "Model1"."id" = 2',
-                    'update "Model1" set "model1Prop1" = \'updated manyToMany 1\' where "Model1"."id" in (select "Model1"."id" from "Model1" inner join "Model1Model2" as "Model1Model2" on "Model1"."id" = "Model1Model2"."model1Id" where "Model1Model2"."model2Id" in (1) and "Model1"."id" = 4 order by "Model1"."id" asc)',
+                    'update "Model1" set "model1Prop1" = \'updated manyToMany 1\' where "Model1"."id" in (select "Model1"."id" from "Model1" inner join "Model1Model2" on "Model1"."id" = "Model1Model2"."model1Id" where "Model1Model2"."model2Id" in (1) and "Model1"."id" = 4 order by "Model1"."id" asc)',
                     'update "model2" set "model2_prop1" = \'updated hasMany 1\', "model1_id" = 2 where "model2"."id_col" = 1 and "model2"."model1_id" in (2)'
                   ]);
               }
@@ -2531,6 +2531,85 @@ module.exports = session => {
                     }
                   });
                 });
+            });
+          });
+      });
+    });
+
+    describe('cyclic references', () => {
+      it('should detect cycles in the graph', done => {
+        const upsert = {
+          id: 1,
+
+          model1Relation1: {
+            '#id': 'root',
+
+            model1Relation1: {
+              '#ref': 'root'
+            }
+          }
+        };
+
+        Model1.bindKnex(session.knex)
+          .query()
+          .upsertGraph(upsert)
+          .then(() => {
+            done(new Error('should not get here'));
+          })
+          .catch(err => {
+            expect(err.message).to.equal('the object graph contains cyclic references');
+            done();
+          })
+          .catch(done);
+      });
+
+      it('cycle detection should consider already inserted nodes', () => {
+        // There are no cycles in this graph because `id=2` has already
+        // been inserted.
+        const upsert = {
+          id: 2,
+
+          model1Relation1: {
+            id: 3,
+            model1Prop1: 'also updated',
+
+            model1Relation1: {
+              '#ref': '@1'
+            }
+          },
+
+          model1Relation1Inverse: {
+            '#id': '@1',
+            model1Prop1: 'hello'
+          }
+        };
+
+        return Model1.bindKnex(session.knex)
+          .query()
+          .upsertGraph(upsert)
+          .then(() => {
+            return Model1.query(session.knex)
+              .findById(2)
+              .eager({
+                model1Relation1: {
+                  model1Relation1: true
+                },
+                model1Relation1Inverse: true
+              });
+          })
+          .then(result => {
+            const id = result.model1Relation1.model1Relation1.id;
+            chai.expect(result).containSubset({
+              id: 2,
+              model1Relation1: {
+                id: 3,
+                model1Relation1: {
+                  id
+                }
+              },
+              model1Relation1Inverse: {
+                id
+              }
             });
           });
       });