+++ typings/objection/index.d.ts
@@ -183,8 +183,23 @@ declare namespace Objection {
     [relationName: string]: RelationMapping;
   }
 
+  interface RelationProperty {
+    size: number;
+    modelClass: ModelClass<any>;
+    props: string[];
+    cols: string[];
+  }
+
   interface Relation {
-    // TODO should this be something other than a tagging interface?
+    name: string;
+    ownerModelClass: ModelClass<any>;
+    relatedModelClass: ModelClass<any>;
+    ownerProp: RelationProperty;
+    relatedProp: RelationProperty;
+    joinModelClass: ModelClass<any>;
+    joinTable: string;
+    joinTableOwnerProp: RelationProperty;
+    joinTableRelatedProp: RelationProperty;
   }
 
   export interface RelationJoin {
@@ -198,6 +213,7 @@ declare namespace Objection {
     to: string | Reference | (string | Reference)[];
     modelClass?: ModelClass<any> | string;
     extra?: string[] | object;
+    beforeInsert?: (model: Model, context: QueryContext) => Promise<void> | void;
   }
 
   export interface RelationMapping {
@@ -222,7 +238,7 @@ declare namespace Objection {
     relate?: boolean | string[];
     unrelate?: boolean | string[];
     insertMissing?: boolean | string[];
-    update?: boolean;
+    update?: boolean | string[];
     noInsert?: boolean | string[];
     noUpdate?: boolean | string[];
     noDelete?: boolean | string[];
@@ -231,7 +247,7 @@ declare namespace Objection {
   }
 
   export interface InsertGraphOptions {
-    relate?: boolean;
+    relate?: boolean | string[];
   }
 
   export interface QueryContext {
@@ -462,6 +478,8 @@ declare namespace Objection {
     static WhereInEagerAlgorithm: EagerAlgorithm;
     static NaiveEagerAlgorithm: EagerAlgorithm;
 
+    static getRelations(): { [key: string]: Relation };
+
     static query<QM extends Model>(
       this: Constructor<QM>,
       trxOrKnex?: Transaction | knex
@@ -636,9 +654,13 @@ declare namespace Objection {
   }
 
   interface InsertGraphAndFetch<QM extends Model> {
-    (modelsOrObjects?: Partial<QM>, options?: InsertGraphOptions): QueryBuilder<QM, QM>;
     (modelsOrObjects?: Partial<QM>[], options?: InsertGraphOptions): QueryBuilder<QM, QM[]>;
+    (modelOrObject?: Partial<QM>, options?: InsertGraphOptions): QueryBuilder<QM, QM>;
   }
+
+  type PartialUpdate<QM extends Model> = {
+    [P in keyof QM]?: QM[P] | Raw | Reference | QueryBuilder<any, any[]>
+  };
 
   interface QueryBuilderBase<QM extends Model, RM, RV> extends QueryInterface<QM, RM, RV> {
     modify(func: (builder: this) => void): this;
@@ -668,16 +690,16 @@ declare namespace Objection {
     /**
      * @return a Promise of the number of updated rows
      */
-    update(modelOrObject: Partial<QM>): QueryBuilderYieldingCount<QM, RM>;
-    updateAndFetch(modelOrObject: Partial<QM>): QueryBuilder<QM, QM>;
-    updateAndFetchById(id: Id, modelOrObject: Partial<QM>): QueryBuilder<QM, QM>;
+    update(modelOrObject: PartialUpdate<QM>): QueryBuilderYieldingCount<QM, RM>;
+    updateAndFetch(modelOrObject: PartialUpdate<QM>): QueryBuilder<QM, QM>;
+    updateAndFetchById(id: Id, modelOrObject: PartialUpdate<QM>): QueryBuilder<QM, QM>;
 
     /**
      * @return a Promise of the number of patched rows
      */
-    patch(modelOrObject: Partial<QM>): QueryBuilderYieldingCount<QM, RM>;
-    patchAndFetchById(idOrIds: IdOrIds, modelOrObject: Partial<QM>): QueryBuilder<QM, QM>;
-    patchAndFetch(modelOrObject: Partial<QM>): QueryBuilder<QM, QM>;
+    patch(modelOrObject: PartialUpdate<QM>): QueryBuilderYieldingCount<QM, RM>;
+    patchAndFetchById(idOrIds: IdOrIds, modelOrObject: PartialUpdate<QM>): QueryBuilder<QM, QM>;
+    patchAndFetch(modelOrObject: PartialUpdate<QM>): QueryBuilder<QM, QM>;
 
     upsertGraph: Upsert<QM>;
     upsertGraphAndFetch: Upsert<QM>;
@@ -1102,6 +1124,7 @@ declare namespace Objection {
       RM,
       RV
     >;
+    (queryBuilder: QueryBuilder<Model>): QueryBuilder<QM, RM, RV>;
   }
 
   interface JoinRaw<QM extends Model, RM, RV> {